refactor: 백엔드 이미지 캐싱과 관광지 API를 개선하고, 프론트엔드 접근성 및 데이터 검증을 강화

+++ backend/src/main/java/com/ssafy/where2meow/attraction/dto/AttractionDetailResponse.java
@@ -19,8 +19,8 @@ public class AttractionDetailResponse {
     private String firstImage1;
     private String firstImage2;
     private Integer mapLevel;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     private String tel;
     private String addr1;
     private String addr2;
